chore: move snackbar-slice to slices
