+++ src/widgets/BhandaraCheckin/store/actions/updateDetailsCheckin/checkinAbhyasi.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { RootState, ThunkApiConfig } from "../..";
-import { snackbarSlice } from "../../../../../components/Snackbar/snackbarSlice";
+import { snackbarSlice } from "../../slices/snackbarSlice";
 import { postAttendance } from "../../api-async-thunks";
 import { UserDetails } from "../../../types";
 import { RefinedCityStateCountryLocation } from "../../../../../components/LocationTextField/locations";
