+++ src/widgets/BhandaraCheckin/store/index.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import { snackbarSlice } from "../../../components/Snackbar/snackbarSlice";
+import { snackbarSlice } from "./slices/snackbarSlice";
 import { BhandaraCheckinAPIs } from "../types";
 import { mainSectionSlice } from "./slices/mainSectionSlice";
 import { updateDetailsSectionSlice } from "./slices/updateDetailsSectionSlice";
