+++ src/widgets/BhandaraCheckin/store/slices/index.ts
@@ -1,1 +1,2 @@
 export * from "./updateDetailsSectionSlice";
+export * from "./snackbarSlice";
